requirements-dev.txt
pytest>=8.2
pytest-django>=4.9
mkdocs>=1.5
mkdocs-material>=9.5
pytest-playwright>=0.5.0
playwright>=1.46
